third_party/blink/web_tests/http/tests/devtools/a11y-axe-core/settings/blackbox-a11y-test-expected.txt

Tests accessibility in Blackboxing view using the axe-core linter.

Running: testAddPattern
Opened input box: true
aXe violations: []


Running: testPatternList
Added a pattern in the list: test*
aXe violations: []


Running: testPatternError
Error message: Pattern cannot be empty
aXe violations: []
